Run beatrix as PID 1 when containerized
Run beatrix with `exec` so that she's PID 1 when running in a container.
This ensures that she'll receive signals (e.g. `docker stop` or hitting
`Ctrl-C` in Compose) and quit immediately; otherwise, the signal gets
eaten by the shell and she ends up being SIGKILL'ed 10 seconds later
(or 30 seconds later on Kubernetes).